Cleaning MainWindow
